requirements.txt
-i https://pypi.org/simple
aiohttp==3.6.3; python_full_version >= '3.5.3'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
chardet==3.0.4
discord.py==1.5.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
multidict==4.7.6; python_version >= '3.5'
pillow==8.0.0
yarl==1.5.1; python_version >= '3.5'